Add command for finding old-style CPs
This adds `ember watson:find-overloaded-cps` for help locating all the
places where your source may trigger the "Using the same function as
getter and setter" deprecation, which is otherwise hard to statically
search for.

Unlike the other commands, this one is purely advisory, it does not
attempt to edit your source.